requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2
django-compat==1.0.15
django4-background-tasks==1.2.7
djangorestframework==3.14.0
fonttools==4.39.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
